// ProjetWeb.Frontend/src/app/core/api/site/model/page-of-of-site-response.ts
/**
 * SiteManagement.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ApiSitesPageGetPageNumberParameter } from './api-sites-page-get-page-number-parameter';
import { SiteResponse } from './site-response';


export interface PageOfOfSiteResponse { 
    pageNumber?: ApiSitesPageGetPageNumberParameter;
    pageSize?: ApiSitesPageGetPageNumberParameter;
    totalItems?: ApiSitesPageGetPageNumberParameter;
    totalPages?: ApiSitesPageGetPageNumberParameter;
    hasPreviousPage?: boolean;
    hasNextPage?: boolean;
    items?: Array<SiteResponse>;
}